+++ src/DataSGen.cpp
@@ -87,14 +87,17 @@ void DataSGen::fillSbus(CplxVect & Sbus, bool ac, const std::vector<int> & id_gr
     int bus_id_me, bus_id_solver;
     cplx_type tmp;
     for(int sgen_id = 0; sgen_id < nb_sgen; ++sgen_id){
-        //  i don't do anything if the load is disconnected
+        //  i don't do anything if the static generator is disconnected
         if(!status_[sgen_id]) continue;
 
         bus_id_me = bus_id_(sgen_id);
         bus_id_solver = id_grid_to_solver[bus_id_me];
         if(bus_id_solver == _deactivated_bus_id){
-            //TODO improve error message with the gen_id
-            throw std::runtime_error("One load is connected to a disconnected bus.");
+            std::ostringstream exc_;
+            exc_ << "DataSGen::fillSbus: Static Generator with id ";
+            exc_ << sgen_id;
+            exc_ << " is connected to a disconnected bus while being connected to the grid.";
+            throw std::runtime_error(exc_.str());
         }
         tmp = static_cast<cplx_type>(p_mw_(sgen_id));
         if(ac) tmp += my_i * q_mvar_(sgen_id);
@@ -125,14 +128,28 @@ void DataSGen::reset_results(){
 void DataSGen::change_p(int sgen_id, real_type new_p, bool & need_reset)
 {
     bool my_status = status_.at(sgen_id); // and this check that load_id is not out of bound
-    if(!my_status) throw std::runtime_error("Impossible to change the active value of a disconnected static generator");
+    if(!my_status)
+    {
+        std::ostringstream exc_;
+        exc_ << "DataSGen::change_p: Impossible to change the active value of a disconnected static generator (check sgen id ";
+        exc_ << sgen_id;
+        exc_ << ")";
+        throw std::runtime_error(exc_.str());
+    }
     p_mw_(sgen_id) = new_p;
 }
 
 void DataSGen::change_q(int sgen_id, real_type new_q, bool & need_reset)
 {
     bool my_status = status_.at(sgen_id); // and this check that load_id is not out of bound
-    if(!my_status) throw std::runtime_error("Impossible to change the reactive value of a disconnected static generator");
+    if(!my_status)
+    {
+        std::ostringstream exc_;
+        exc_ << "DataSGen::change_q: Impossible to change the reactive value of a disconnected static generator (check sgen id ";
+        exc_ << sgen_id;
+        exc_ << ")";
+        throw std::runtime_error(exc_.str());
+    }
     q_mvar_(sgen_id) = new_q;
 }