+++ src/DataShunt.h
@@ -9,12 +9,14 @@
 #ifndef DATASHUNT_H
 #define DATASHUNT_H
 
+#include "Utils.h"
+
 #include "Eigen/Core"
 #include "Eigen/Dense"
 #include "Eigen/SparseCore"
 #include "Eigen/SparseLU"
 
-#include "Utils.h"
+
 #include "DataGeneric.h"
 
 /**